+++ changelog.md
@@ -1,3 +1,6 @@
 - fixed translation of some chat messages
 - added sound when placing pattern
 - fix version number
+- add <hold shift> to dictionary tooltip
+- add JEI description page to dictionary, staff, broom and pouch
+- fix RecipeSorter's issues with the pouch